Added prune func for winningBoardStates that I'm not sure is necessary
I added a function that would "prune" the winningBoardStates down to only the
first win that results from a move in a given cell.  At this point I'm not
sure it's necessary for the minimax calculation to return the optimum next
move.  I do need to re-work the returnComputerMove function to disregard cells
that have already been played.